pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fremorizer"
description = "Model output rewriter (CMORizer) for FRE/FMS based models"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    {name = "MSD Workflow Team", email="oar.gfdl.workflow@noaa.gov"}
]

keywords = [
    "gfdl",
    "workflow",
    "fre",
    "fms",
    "cmor",
]

dependencies = [
        'cftime',
        'click',
        'cmor',
        'jsonschema',
        'netCDF4',
        'numpy',
        'pyyaml',
]

classifiers = [
    "Environment :: Console",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
]

dynamic = [
    "version",
]


[tool.setuptools.dynamic]
version = {attr = "fremorizer._version.__version__"}


[project.scripts]
fremor = 'fremorizer.cli:fremor'


[project.optional-dependencies]
docs = [
    "sphinx",
    "renku-sphinx-theme",
    "sphinx-rtd-theme",
    "rstcheck",
]

test = [
    "pylint",
    "pytest",
]

qa = [
    "cc-plugin-wcrp",
]


[tool.rstcheck]
report_level = "WARNING"
ignore_directives = [
    "automodule",
    "autoclass",
    "autofunction",
    "automethod",
    "autoattribute",
    "toctree",
    "glossary",
]
ignore_roles = ["ref"]
